requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
slowapi>=0.1.9
dnspython>=2.7.0
phonenumbers>=8.13.0
qrcode[pil]>=8.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.12
pillow>=10.0.0
validators>=0.34.0
